+++ src/Router.tsx
@@ -5,9 +5,9 @@ import SignUp from "./pages/SignUp";
 import Login from "@/pages/Login";
 import Reservation from "@/pages/Reservation";
 import MatchList from "@/pages/MatchList";
-import Admin from "@/pages/admin/index";
+import Admin from "@/pages/Admin/index";
 import NotFound from "./pages/NotFound";
-import Registration from "@/pages/admin/Registration/index";
+import Registration from "@/pages/Registration/index";
 import { menu } from "./common/constants";
 import AuthRoute from "./common/components/AuthLayout";
 import AccountDeletion from "@/pages/AccountDeletion";
