+++ src/pages/Registration/components/SelectField.tsx
@@ -1,5 +1,5 @@
 import InputErrorMessage from "@/common/components/atoms/InputMessage";
-import { FormValueType } from "@/pages/admin/type";
+import { FormValueType } from "@/pages/Admin/type";
 import { FieldError, UseFormRegister } from "react-hook-form";
 
 interface SelectProps {
