+++ src/pages/Registration/hooks/useRegistrationForm.ts
@@ -2,8 +2,8 @@ import { useNavigate } from "react-router-dom";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import useStore from "@/common/stores/useStore";
-import { FormValueType } from "@/pages/admin/type";
-import { regiSchema } from "@/pages/admin/Registration/validation/RegiSchema";
+import { FormValueType } from "@/pages/Admin/type";
+import { regiSchema } from "@/pages/Registration/validation/RegiSchema";
 import { getTeamId } from "@/common/utils/getTeamId";
 import alertToast from "@/common/utils/alertToast";
 import { MainMatchType } from "@/common/types/matchTypes";
