+++ src/pages/admin/components/DetailModal/DetailInfo.tsx
