+++ src/pages/admin/components/DetailModal/index.tsx
@@ -4,7 +4,7 @@ import { MainMatchType } from "@/common/types/matchTypes";
 import alertToast from "@/common/utils/alertToast";
 import ErrorToast from "@/common/utils/errorToast";
 import confirmAlert from "@/common/utils/confirmAlert";
-import DetailInfo from "./components/DetailInfo";
+import DetailInfo from "./DetailInfo";
 import axiosInstance from "@/common/utils/axiosInstance";
 
 interface DetailModalProps {
